Don't force latest k6 when building extensions
an extension will always require k6 and a version of it. It makes little
sense to force newer version on users especially as we keep breaking
them and we need time to go fix them, but that can only happen after
release.

This still lets users specify the version (as latest for example) and
require the latest k6, it just isn't the default.