Fix any kind of go module replacements getting latest k6
This is a bit strange but it turns out that if you have replacements go
change how it resolves modules a bit and will try to get the latest
versions of some dependencies.

Also if you have one extension and you are building it from local
sources we would not run `go mod tidy` reliably. This doesn't really
matter in practice as we do run it almost every other command, but was
getting in the way of smaller fix.

Both of those are fixed with this commit and now we do definitely run
`go mod tidy` once we have requested all extensions and put down the
files needing them. And also only write a file requesting k6 directly at
the very end so it only try to get it's version at this point - and at
that point we should already have requested through an extension.