CI: Add build for linux/arm64
This builds and pushes a multi-arch Docker manifest for amd64 and arm64.

We use Docker's provided action to do this using qemu emulation, and
switch to using this to push the images for us.

Split the build/test of the `xk6` binary out to a separate job too, so
that can also be run on arm64.